+++ src/App.vue
@@ -1,9 +1,16 @@
 <template>
 
-
-  <!-- <BarSide :menus="menus" /> -->
-
+ 
+  <h4> CalendarMonthly </h4>
+  <CalendarMonthly/>
+    <CalendarWeekly/>
   <CardExample />
+
+ <!--  <h4> CalendarWeekly </h4>
+
+  
+  <BarSide :menus="menus" /> 
+
   <BadgeExample />
 
   <AccordionExemple />
@@ -20,12 +27,13 @@
     @unsort="unsort" 
     @sort-asc="sortAsc"
     @sort-desc="sortDesc"
-    @hide-column="hideColumn"/>
+    @hide-column="hideColumn"/>-->
+
 </template>
 
 <script>
 
-import BarSideExemple from './components/barside/BarSideExemple.vue';
+/*import BarSideExemple from './components/barside/BarSideExemple.vue';
 import InputExemple from './components/input/InputExemple.vue';
 import ButtonExample from './components/button/ButtonExample.vue';
 
@@ -38,12 +46,18 @@ import RadioExemple from './components/radio/RadioExemple.vue';
 import CheckBoxExemple from './components/checkbox/CheckBoxExemple.vue';
 import AccordionExemple from './components/accordion/AccordionExemple.vue';
 
-import BadgeExample from './components/badge/BadgeExample.vue';
+import BadgeExample from './components/badge/BadgeExample.vue';*/
+import CalendarWeekly from './components/calendar/CalendarWeekly.vue';
+import CalendarMonthly from './components/calendar/month/CalendarMonthly.vue';
 import CardExample from './components/card/CardExample.vue';
+
 
 export default {
 
   components: {
+    
+    
+    /*
     BarSideExemple,
     InputExemple,
     DataTable,
@@ -53,8 +67,11 @@ export default {
     RadioExemple,
     CheckBoxExemple,
     AccordionExemple,
-    BadgeExample,
-    CardExample
+    BadgeExample
+    BadgeExample,*/
+    CalendarMonthly,
+    CalendarWeekly,
+    CardExample,
 
 
   },
@@ -62,7 +79,7 @@ export default {
   data() {
     return {
 
-      menus: [
+     /*menus: [
         [
           {
             title: "Dashboard",
@@ -275,7 +292,7 @@ export default {
         pageStart: 1,
         totalRecords: 50,
         pageLength :[10, 20, 25, 50, 100]
-      },
+      },*/
     };
   },