+++ src/assets/icons/MenuIcon.vue
@@ -1,0 +1,37 @@
+<template>
+
+	<SvgIcon v-bind="this.$props">
+		<svg width="28" height="29" viewBox="0 0 28 29" fill="none" xmlns="http://www.w3.org/2000/svg">
+			<rect y="0.516602" width="28" height="28" rx="14" fill="white"/>
+			<g clip-path="url(#clip0_610_14914)">
+			<path d="M14 15.2666C14.4142 15.2666 14.75 14.9308 14.75 14.5166C14.75 14.1024 14.4142 13.7666 14 13.7666C13.5858 13.7666 13.25 14.1024 13.25 14.5166C13.25 14.9308 13.5858 15.2666 14 15.2666Z" stroke="#002A34" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+			<path d="M14 20.5166C14.4142 20.5166 14.75 20.1808 14.75 19.7666C14.75 19.3524 14.4142 19.0166 14 19.0166C13.5858 19.0166 13.25 19.3524 13.25 19.7666C13.25 20.1808 13.5858 20.5166 14 20.5166Z" stroke="#002A34" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+			<path d="M14 10.0166C14.4142 10.0166 14.75 9.68082 14.75 9.2666C14.75 8.85239 14.4142 8.5166 14 8.5166C13.5858 8.5166 13.25 8.85239 13.25 9.2666C13.25 9.68082 13.5858 10.0166 14 10.0166Z" stroke="#002A34" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+			</g>
+			<defs>
+			<clipPath id="clip0_610_14914">
+			<rect width="18" height="18" fill="white" transform="translate(5 5.5166)"/>
+			</clipPath>
+			</defs>
+			</svg>
+
+	</SvgIcon>
+
+</template>
+
+<script>
+import SvgIcon from './SvgIcon.vue';
+
+
+export default {
+
+	name: 'MenuIcon',
+	components: {
+		SvgIcon
+	}
+
+}
+
+</script>
+
+<style lang="scss"></style>