+++ src/assets/icons/SvgIcon.vue
@@ -1,0 +1,33 @@
+<template>
+
+    <span :class="className+' svg-ico'">
+        <slot></slot>
+    </span>
+
+</template>
+
+<script>
+
+
+export default {
+
+    name: 'SvgIcon',
+    props: {
+        className : {
+            type: String,
+            required : false,
+        }
+    }
+
+}
+
+</script>
+
+<style lang="scss">
+
+.svg-ico {
+    display: flex;
+    align-items: center;
+}
+
+</style>