+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Created By: BENGUERGOURA Oussama & HADJERSI Mohamed | 20/10/2024
 
 ### Added
 
+* Adding `past, today, future` css class to handle display. [GTS-PRINT#35] (https://quire.io/w/GTS-PRINT31/35)
+* Adding `CalendarMonthlyInfo` component to handle calendar day information. [GTS-PRINT#35] (https://quire.io/w/GTS-PRINT31/35)
+* Adding First part of `CalendarMonthly` component. [GTS-PRINT#35] (https://quire.io/w/GTS-PRINT31/35)
+* Adding First part of `CalendarWeekly` component. [GTS-PRINT#35] (https://quire.io/w/GTS-PRINT31/35)
 * Adding `CardComponent` with test and documentation. [GTS-PRINT#32](https://quire.io/w/GTS-PRINT31/32)
 * Adding documentation for `BadgeComponent`. [GTS-PRINT#30](https://quire.io/w/GTS-PRINT31/30)
 * Adding `badges.js` file which contain constant badge css theme class [GTS-PRINT#30](https://quire.io/w/GTS-PRINT31/30)